+++ src/main/java/com/dalbit/team/vo/TeamResultVo.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter @Setter
 public class TeamResultVo {
    Integer team_no;			    //BIGINT		-- 팀번호
-   Integer master_mem_no;		//BIGINT		-- 팀장회원번호
+   Long master_mem_no;		//BIGINT		-- 팀장회원번호
    String team_name;		        //VARCHAR(15)	-- 팀명
    String team_conts;		    //VARCHAR(200)	-- 팀소개
    String team_medal_code;		//CHAR(4)		-- 팀메달
@@ -23,7 +23,7 @@ public class TeamResultVo {
    String req_mem_yn;		    //CHAR(1)		-- 가입신청 허용여부
    String ins_date;		        //DATETIME	-- 등록일자
    String upd_date;		        //DATETIME	-- 수정일자
-   Integer tm_mem_no;		    //BIGINT		-- 팀원회원번호
+   Long tm_mem_no;		    //BIGINT		-- 팀원회원번호
    String tm_mem_nick;		    //VARCHAR(20)	-- 팀원대화명
    String tm_image_background;	//VARCHAR(256)	-- 프로필 배경사진
    Integer tm_mem_score;		    //BIGINT		-- 팀 기여점수
